requirements.txt
django-phonenumber-field==7.0.2
phonenumbers==8.13.4
django==3.2.6
Pillow==9.2.0
djangorestframework==3.13.1
virtualenv==20.17.1
distlib==0.3.6
filelock==3.9.0
platformdirs==2.6.2
django-environ==0.9.0
psycopg2-binary==2.9.5
djangorestframework-simplejwt==5.2.2
pyjwt==2.6.0
cffi==1.15.1
cryptography~=38.0.4
pycparser==2.21
defusedxml~=0.7.1
future~=0.18.2
pip~=22.3.1
toml~=0.10.2
py~=1.10.0
lxml~=4.6.4
api~=0.0.7
asyncio~=3.4.3
pytz~=2016.4
sqlparse~=0.4.2
psycopg2~=2.9.3
asgiref~=3.5.2
colorama~=0.4.5
PyYAML~=5.4.1
Jinja2~=3.1.2
numpy~=1.21.3
serializers~=0.2.4
simplejson~=3.17.6
environ~=1.0
ply~=3.11
setuptools~=57.4.0
Cython~=0.29.30
python2~=1.2
bcrypt~=3.2.2
tzdata~=2022.1
coreapi~=2.3.3
uritemplate~=4.1.1
coreschema~=0.0.4
requests~=2.26.0
Markdown~=3.4.1
Pygments~=2.12.0
Babel~=2.9.1
channels~=3.0.4
loguru~=0.5.3
django-filter~=22.1